taskvine py: setup into a directory
Also, clearly divide what comes from C.
Remove result in favor of result_string (which was renamed result).

C constants renamed as:
VINE_ALLOCATION_Y -> "y"
