+++ taskvine/src/bindings/python3/taskvine/display.py
@@ -1,3 +1,7 @@
+# Copyright (C) 2022- The University of Notre Dame
+# This software is distributed under the GNU General Public License.
+# See the file COPYING for details.
+
 import numbers
 import time
 import sys
