+++ taskvine/test/vine_python_task.py
@@ -10,17 +10,20 @@ except IndexError:
     sys.stderr.write("Usage: {} PORTFILE\n".format(sys.argv[0]))
     raise
 
+
 # Define a function to invoke remotely
 def my_sum(x, y, negate=False):
-    import math
+    from operator import add, mul
+
     f = 1
     if negate:
         f = -1
-    s = f * (x + y)
+    s = mul(f, add(x, y))
     return s
 
+
 # Create a new queue
-queue = vine.Manager(port=[9123,9130])
+queue = vine.Manager(port=[9123, 9130])
 print("listening on port {}".format(queue.port))
 with open(port_file, "w") as f:
     f.write(str(queue.port))
@@ -28,7 +31,7 @@ with open(port_file, "w") as f:
 
 # Submit several tasks for execution:
 print("submitting tasks...")
-for value in range(1,10):
+for value in range(1, 10):
     task = vine.PythonTask(my_sum, value, value)
     task.set_cores(1)
     queue.submit(task)
@@ -38,12 +41,12 @@ positive_sum = 0
 while not queue.empty():
     task = queue.wait(5)
     if task:
-        print("task {} completed with result {}".format(task.id,task.output))
+        print("task {} completed with result '{}': {}".format(task.id, task.result, task.output))
         positive_sum += task.output
 
 
 # Submit several tasks for execution:
-for value in range(1,10):
+for value in range(1, 10):
     task = vine.PythonTask(my_sum, value, value, negate=True)
     task.set_cores(1)
     queue.submit(task)
@@ -53,9 +56,7 @@ negative_sum = 0
 while not queue.empty():
     task = queue.wait(5)
     if task:
-        print("task {} completed with result {}".format(task.id,task.output))
+        print("task {} completed with result '{}': {}".format(task.id, task.result, task.output))
         negative_sum += task.output
 
-
-assert(positive_sum == (-1 * negative_sum))
-
+assert positive_sum == (-1 * negative_sum)
