+++ taskvine/test/vine_test_tag.py
@@ -18,8 +18,8 @@ alpha_order = sorted(desired_tag_order)
 done_order = []
 
 q = vine.Manager(port=0)
-with open(port_file, 'w') as f:
-    print('Writing port {port} to file {file}'.format(port=q.port, file=port_file))
+with open(port_file, "w") as f:
+    print("Writing port {port} to file {file}".format(port=q.port, file=port_file))
     f.write(str(q.port))
 print(vine.__file__)
